+++ src/app/config/static.service.ts
@@ -21,4 +21,17 @@ export class StaticService {
   getKIConfig(): any {
     return CustomConfig.kiConfig;
   }
+  getKIModelPathMNIST(): any {
+    return CustomConfig.kiConfig.mnistPath;
+  }
+  getKIModelPathSentiment(lang: string = 'en'): any {
+    if (lang=='de') {
+      return CustomConfig.kiConfig.sentimentPath.de;
+    } else{
+      return CustomConfig.kiConfig.sentimentPath.en;
+    }
+  }
+  getCourseNumbers(): any {
+    return CustomConfig.courseNumbers;
+  }
 }