+++ src/app/sites/ki-tools/ki-tools-routing.module.ts
@@ -1,7 +1,9 @@
-import { MnistComponent } from './pages/mnist/mnist-standalone/mnist.component';
+import { DemonstratorsStandaloneComponent } from './pages/demonstrators/demonstrators-standalone.component';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { KIPlaygroundComponent } from 'src/app/sites/ki-tools/pages/ki-playground/ki-playground.component';
+import { MnistStandaloneComponent } from './pages/mnist/mnist-standalone.component';
+import { SentimentStandAloneComponent } from 'src/app/sites/ki-tools/pages/sentiment/sentiment-standalone.component';
 
 const routes: Routes = [
   {
@@ -10,7 +12,16 @@ const routes: Routes = [
   },
   {
     path: 'mnist',
-    component: MnistComponent
+    component: MnistStandaloneComponent
+  }
+  ,
+  {
+    path: 'sentiment',
+    component: SentimentStandAloneComponent
+  },
+  {
+    path: 'examples',
+    component: DemonstratorsStandaloneComponent
   }
 ];
 @NgModule({