+++ src/app/sites/ki-tools/ki-tools.module.ts
@@ -1,27 +1,42 @@
 import { NgModule } from '@angular/core';
 import { SharedModule } from 'src/app/shared/shared.module';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { KiToolsRoutingModule } from './ki-tools-routing.module';
 import { KiStatusService } from './services/ki-status.service';
 import { MNISTExampleComponent } from './components/mnist-example/mnist-example.component';
 import { KIPlaygroundComponent } from './pages/ki-playground/ki-playground.component';
 import { DrawableCanvasComponent } from './components/mnist-example/drawable-canvas/drawable-canvas.component';
-import { NgbdMnistModalComponent } from './pages/mnist/mnist-modal/mnist-modal.component';
-import { MnistComponent } from './pages/mnist/mnist-standalone/mnist.component';
-import { MnistLinkFooterComponent } from './components/mnist-example/mnist-link-footer/mnist-link-footer.component';
+import { NgbdMnistModalComponent } from './pages/mnist/mnist-modal.component';
+import { MnistStandaloneComponent } from './pages/mnist/mnist-standalone.component';
 import { EventCanvasDirective } from './components/mnist-example/drawable-canvas/event-canvas.directive';
+import { SentimentStandAloneComponent } from './pages/sentiment/sentiment-standalone.component';
+import { NgbdSentimentModalComponent } from './pages/sentiment/sentiment-modal.component';
+import { SentimentExampleComponent } from './components/sentiment-example/sentiment-example.component';
+import { DemonstratorsStandaloneComponent } from './pages/demonstrators/demonstrators-standalone.component';
+import { NgbdDemonstratorsModalComponent } from './pages/demonstrators/demonstrators-modal.component';
+import { ExampleCardComponent } from './components/demonstrator-example/example-card/example-card.component';
+import { DemonstratorExampleComponent } from './components/demonstrator-example/demonstrator-example.component';
+import { CourseListComponent } from './components/shared/course-list/course-list.component';
 
 @NgModule({
   declarations: [
     MNISTExampleComponent,
     KIPlaygroundComponent,
-    MnistComponent,
+    MnistStandaloneComponent,
     DrawableCanvasComponent,
     NgbdMnistModalComponent,
-    MnistLinkFooterComponent,
-    EventCanvasDirective
+    EventCanvasDirective,
+    SentimentStandAloneComponent,
+    NgbdSentimentModalComponent,
+    SentimentExampleComponent,
+    DemonstratorsStandaloneComponent,
+    NgbdDemonstratorsModalComponent,
+    ExampleCardComponent,
+    DemonstratorExampleComponent,
+    CourseListComponent,
   ],
-  imports: [SharedModule, KiToolsRoutingModule],
+  imports: [SharedModule, KiToolsRoutingModule, ReactiveFormsModule, FormsModule],
   providers: [KiStatusService]
 })
 export class KiToolsModule {}