+++ src/app/sites/ki-tools/services/ki-status.service.ts
@@ -2,7 +2,8 @@ import { Injectable, Renderer2 } from '@angular/core';
 import { ScriptLoaderService } from 'src/app/core/services/script-loader/script-loader.service';
 import { StaticService } from 'src/app/config/static.service';
 import { environment } from 'src/environments/environment';
-import { AsyncSubject, Observable, BehaviorSubject, from, throwError } from 'rxjs';
+import { AsyncSubject, Observable, from, throwError } from 'rxjs';
+import { HttpClient } from '@angular/common/http';
 
 declare var tf: any;
 
@@ -10,11 +11,8 @@ declare var tf: any;
 export class KiStatusService {
   private scriptLoading$: AsyncSubject<any>;
   private MNISTModel$: AsyncSubject<any>;
-
-  private kiToolsModelPath = environment.modelURL + this.staticService.getKIConfig().mnistPath;
-  //public imageDraw: BehaviorSubject<any> = new BehaviorSubject(null);
-
-  constructor(private scriptLoader: ScriptLoaderService, private staticService: StaticService) {}
+  private SentimentModel$: AsyncSubject<any>;
+  constructor(private scriptLoader: ScriptLoaderService, private staticService: StaticService, private httpClient: HttpClient) {}
 
   public loadKIScript(renderer: Renderer2): Observable<any> {
     return new Observable((observer$) => {
@@ -28,17 +26,39 @@ export class KiStatusService {
   }
 
   public loadMNISTModel(){
+  const  kiToolsMnistModelPath = environment.modelURL + this.staticService.getKIModelPathMNIST() + '/model.json';
     if (this.scriptLoading$) {
       return new Observable((observer$) => {
         if (!this.MNISTModel$) {
-         console.log("Load Model");
+         console.log("Load MNIST-Model:", kiToolsMnistModelPath);
          this.MNISTModel$ = new AsyncSubject();
-         from(tf.loadLayersModel(this.kiToolsModelPath)).subscribe(this.MNISTModel$);
+         from(tf.loadLayersModel(kiToolsMnistModelPath)).subscribe(this.MNISTModel$);
         }
         return this.MNISTModel$.subscribe(observer$);
       });
     } else {
       throw throwError('Tensorflow was not loaded');
     }
+  }
+
+  public loadSentimentModel(lang: string = 'en'): Observable<any>{
+    const kiToolsSentimentModelPath_en = environment.modelURL + this.staticService.getKIModelPathSentiment(lang) + '/model/model.json';
+    if (this.scriptLoading$) {
+      return new Observable((observer$) => {
+        if (!this.SentimentModel$) {
+         console.log("Load Sentiment-Model");
+         this.SentimentModel$ = new AsyncSubject();
+         from(tf.loadLayersModel(kiToolsSentimentModelPath_en)).subscribe(this.SentimentModel$);
+        }
+        return this.SentimentModel$.subscribe(observer$);
+      });
+    } else {
+      throw throwError('Tensorflow was not loaded');
+    }
+  }
+
+  public loadWordIndex(lang: string = 'en'): Observable<any>{
+    const kiToolsSentimentIndexPath_en = environment.modelURL + this.staticService.getKIModelPathSentiment(lang) + '/imdb_word_index.json';
+    return this.httpClient.get(kiToolsSentimentIndexPath_en);
   }
 }